Rename filename to fit PEP8
