+++ dapr/actor/runtime/method_context.py
@@ -5,7 +5,7 @@ Copyright (c) Microsoft Corporation.
 Licensed under the MIT License.
 """
 
-from dapr.actor.runtime.calltype import ActorCallType
+from dapr.actor.runtime.call_type import ActorCallType
 
 
 class ActorMethodContext:
