+++ dapr/actor/runtime/runtime.py
@@ -11,7 +11,7 @@ from dapr.actor.id import ActorId
 from dapr.actor.runtime.actor import Actor
 from dapr.actor.runtime.config import ActorRuntimeConfig
 from dapr.actor.runtime.context import ActorRuntimeContext
-from dapr.actor.runtime.typeinformation import ActorTypeInformation
+from dapr.actor.runtime.type_information import ActorTypeInformation
 from dapr.actor.runtime.manager import ActorManager
 from dapr.clients import DaprActorHttpClient
 from dapr.serializers import Serializer, DefaultJSONSerializer
