+++ tests/actor/test_actor.py
@@ -13,7 +13,7 @@ from unittest.mock import AsyncMock
 from dapr.actor.runtime.config import ActorRuntimeConfig
 from dapr.actor.runtime.context import ActorRuntimeContext
 from dapr.actor.runtime.runtime import ActorRuntime
-from dapr.actor.runtime.typeinformation import ActorTypeInformation
+from dapr.actor.runtime.type_information import ActorTypeInformation
 from dapr.serializers import DefaultJSONSerializer
 
 from tests.actor.fake_actor_classes import (
