+++ tests/actor/test_method_dispatcher.py
@@ -9,8 +9,8 @@ import unittest
 from unittest.mock import AsyncMock
 
 from dapr.actor.runtime.context import ActorRuntimeContext
-from dapr.actor.runtime.methoddispatcher import ActorMethodDispatcher
-from dapr.actor.runtime.typeinformation import ActorTypeInformation
+from dapr.actor.runtime.method_dispatcher import ActorMethodDispatcher
+from dapr.actor.runtime.type_information import ActorTypeInformation
 from dapr.serializers import DefaultJSONSerializer
 
 from tests.actor.fake_actor_classes import FakeSimpleActor
