+++ tests/actor/test_reminder_data.py
@@ -8,7 +8,7 @@ Licensed under the MIT License.
 import unittest
 from datetime import timedelta
 
-from dapr.actor.runtime.reminderdata import ActorReminderData
+from dapr.actor.runtime.reminder_data import ActorReminderData
 
 
 class ActorReminderTests(unittest.TestCase):
