+++ tests/actor/test_type_information.py
@@ -7,7 +7,7 @@ Licensed under the MIT License.
 
 import unittest
 
-from dapr.actor.runtime.typeinformation import ActorTypeInformation
+from dapr.actor.runtime.type_information import ActorTypeInformation
 from tests.actor.fake_actor_classes import (
     FakeSimpleActor,
     FakeMultiInterfacesActor,
